Time: 19 ms (80.29%), Space: 17.4 MB (83.63%) - LeetHub

+++ 0018-4sum/0018-4sum.cpp
@@ -55,7 +55,7 @@ public:
                 if(j != i+1 && nums[j]==nums[j-1])continue;
                 int k = j + 1, l = n - 1;
                 while (k < l) {
-                    long long sum = nums[i] + nums[j] + nums[k] + nums[l];
+                    long long sum = (long long)nums[i] + nums[j] + nums[k] + nums[l];
                     if (sum < target) {
                         k++;
                     } else if (sum > target) {
